requirements.txt
pytorch-lightning==1.4.1
torch==1.6.0
torchvision==0.7.0
captcha
pyyaml
pillow
tqdm